feat(3023): process jobs partners: bloque les offres dans la liste cfa
