+++ server/src/jobs/offrePartenaire/isCompanyInBlockedCfaList.ts
@@ -1,0 +1,1788 @@
+import { removeAccents } from "shared"
+
+const cfaCompanyList = [
+  "ISCOM Lille",
+  "CFA IESA PARIS",
+  "CFA IESA STRASBOURG",
+  "GROUPE IMT",
+  "PERSPECTIVIA",
+  "ATHENA BUSINESS SCHOOL",
+  "INHNI",
+  "INHNI VILLEJUIF",
+  "INHNI VILLEJUIF II",
+  "INHNI GENNEVILLIERS",
+  "INHNI SCHILTIGHEIM",
+  "INHNI VILLENEUVE-D’ASCQ",
+  "INHNI METZ",
+  "INHNI SAINT-HERBLAIN",
+  "INHNI ANGERS",
+  "INHNI POITIERS",
+  "INHNI TOURS",
+  "INHNI MARSEILLE",
+  "INHNI NOUVELLE AQUITAINE",
+  "INHNI PESSAC",
+  "INHNI BRUZ",
+  "INHNI TOULOUSE",
+  "INHNI VENISSIEUX",
+  "ACADELY",
+  "STUDENCY",
+  "TETRANERGY BUSINESS SCHOOL REUNION",
+  "PIGIER AIX-EN-PROVENCE",
+  "PIGIER AMIENS",
+  "PIGIER BAYONNE",
+  "PIGIER BESANÇON",
+  "PIGIER BORDEAUX",
+  "PIGIER CLERMONT-FERRAND",
+  "PIGIER DIJON",
+  "PIGIER GRENOBLE",
+  "PIGIER LENS",
+  "PIGIER LILLE",
+  "PIGIER LORIENT",
+  "PIGIER LYON",
+  "PIGIER MARSEILLE",
+  "PIGIER MELUN",
+  "PIGIER METZ",
+  "PIGIER MONTPELLIER",
+  "PIGIER MULHOUSE",
+  "PIGIER NANCY",
+  "PIGIER NANTES",
+  "PIGIER NICE",
+  "PIGIER NÎMES",
+  "PIGIER ORLÉANS",
+  "PIGIER PARIS",
+  "PIGIER REIMS",
+  "PIGIER ROUEN",
+  "PIGIER SAINT-QUENTIN-EN-YVELINES",
+  "PIGIER STRASBOURG",
+  "PIGIER TOULOUSE",
+  "PIGIER TOURS",
+  "PIGIER TROYES",
+  "PIGIER VALENCIENNES",
+  "MYDIGITALSCHOOL ANGERS",
+  "MYDIGITALSCHOOL ANNECY",
+  "MYDIGITALSCHOOL BORDEAUX",
+  "MYDIGITALSCHOOL CAEN",
+  "MYDIGITALSCHOOL GRENOBLE",
+  "MYDIGITALSCHOOL LILLE",
+  "MYDIGITALSCHOOL LYON",
+  "MYDIGITALSCHOOL MELUN",
+  "MYDIGITALSCHOOL MONTPELLIER",
+  "MYDIGITALSCHOOL NANCY",
+  "MYDIGITALSCHOOL NANTES",
+  "MYDIGITALSCHOOL NICE",
+  "MYDIGITALSCHOOL PARIS",
+  "MYDIGITALSCHOOL RENNES",
+  "MYDIGITALSCHOOL SAINT-QUENTIN-EN-YVELINES",
+  "MYDIGITALSCHOOL TOULOUSE",
+  "MYDIGITALSCHOOL VANNES",
+  "MBWAY ANGERS",
+  "MBWAY ANNECY",
+  "MBWAY BORDEAUX",
+  "MBWAY BREST",
+  "MBWAY CAEN",
+  "MBWAY CHAMBÉRY",
+  "MBWAY GRAND GENÈVE",
+  "MBWAY GRENOBLE",
+  "MBWAY LAVAL",
+  "MBWAY LILLE",
+  "MBWAY LYON",
+  "MBWAY MONTPELLIER",
+  "MBWAY NANTES",
+  "MBWAY PARIS OUEST",
+  "MBWAY PARIS SUD",
+  "MBWAY RENNES",
+  "MBWAY STRASBOURG",
+  "MBWAY TOULOUSE",
+  "MBWAY VANNES",
+  "ESAP BORDEAUX",
+  "ESAP LILLE",
+  "ESAP LYON",
+  "ESAP PARIS",
+  "ESAP TOULOUSE",
+  "WIN ALBERTVILLE",
+  "WIN ANGERS",
+  "WIN ANNECY",
+  "WIN BAYONNE",
+  "WIN BORDEAUX",
+  "WIN CAEN",
+  "WIN CHAMBERY",
+  "WIN GENÈVE",
+  "WIN LAVAL",
+  "WIN LILLE",
+  "WIN LYON",
+  "WIN MADRID",
+  "WIN MARSEILLE",
+  "WIN MELUN",
+  "WIN MONTPELLIER",
+  "WIN NANCY",
+  "WIN NANTES",
+  "WIN NICE",
+  "WIN PARIS",
+  "WIN REIMS",
+  "WIN RENNES",
+  "WIN SAINT-QUENTIN-EN-YVELINES",
+  "WIN STRASBOURG",
+  "WIN TOULOUSE",
+  "WIN TOURS",
+  "WIN VANNES",
+  "Afpa pays de Savoie",
+  "AFTEC Caen",
+  "iscod alternance",
+  "AFTEC Rennes",
+  "IFAE",
+  "AFTEC Vannes",
+  "AFTEC Brest",
+  "Tetranergy Business School Réunion",
+  "Tetranergy Business School Rodez",
+  "CFA HORIZON REUNION BUSINESS SCHOOL",
+  "OUTREMER ACADEMY",
+  "GROUPE ENDEMIA FORMATION",
+  "ISFOI",
+  "PÔLE FORMATION UIMM BRETAGNE - BRUZ",
+  "PÔLE FORMATION UIMM BRETAGNE - PLERIN",
+  "IFAC Campus des Métiers BREST",
+  "IFCV - Formations en Alternance",
+  "IPAC Bachelor Factory Bordeaux",
+  "IGENSIA Alternance",
+  "IPAC Bachelor Factory Caen",
+  "IPAC Bachelor Factory Montpellier",
+  "AFTEC Laval",
+  "CFAO",
+  "EFREI - Grande école du numérique",
+  "INA - Institut national de l'audiovisuel",
+  "IPAC Bachelor Factory Angers",
+  "IPAC Bachelor Factory Vannes",
+  "ÉCOLES AFTEC",
+  "CFA DIFCAM",
+  "IFCAM - Université du Groupe Crédit Agricole",
+  "École de Management Appliqué (EMA)",
+  "CESI",
+  "ESCCOT GROUPE Ecole de Commerce et CFA",
+  "ICANS | Institut de cancérologie Strasbourg Europe",
+  "IPAC Albertville",
+  "IPAC Bachelor Factory Albertville",
+  "Institut Européen de Formation",
+  "Ipac Bachelor Factory",
+  "OMNES Education",
+  "Turnpoint Executive Search | Supply Chain, Logistics & Transport",
+  "École Intuit Lab",
+  "IFCDis",
+  "Chambre Régionale de Métiers et de l'Artisanat",
+  "CIO'SUP Drôme-Ardèche",
+  "Pole Formation Auvergne / CFAI",
+  "FORMATION AGENT IMMOBILIER",
+  "GROUPE DE FORMATION",
+  "GFS - Grande École de l'Alternance",
+  "Institut Mines-Télécom Business School",
+  "13 EN FORM",
+  "1PULSE FORMATION",
+  "2 CG",
+  "2CRD",
+  "3.14 ACADEMY",
+  "301",
+  "3IE",
+  "3T DEVELOPPEMENT",
+  "A E P ENTRE D'AFFAIRES",
+  "A F P C NORD PAS DE CALAIS",
+  "A F P O L S",
+  "A M B FORMATIONS",
+  "A M I G R A F",
+  "AMIT",
+  "A ZOT FORMATION",
+  "A.J.M",
+  "A.MI.PY",
+  "ABC CONSEIL ET FORMATION",
+  "ABL FORMATION",
+  "ABSUP",
+  "ACADEMIE D'ART DENTAIRE",
+  "ACADEMIE DE LA BEAUTE",
+  "ACADEMIE DES GOUVERNANTES",
+  "ACADÉMIE DU TOURISME",
+  "ACADEMIE FORMATION GROUPE COMPASS",
+  "ACADEMIE NATIONALE DE LA BOULANGERIE PATISSERIE",
+  "ACADEMIE.LAX FORMATION PROFESSIONNELLE",
+  "ACAF-MSA DE VAUCLUSE",
+  "ACCECE",
+  "ACCOMPAGNEMENT ET FORMATION POUR LE DEVELOPPEMENT DES ENTREPRISES",
+  "ACCOMPAGNEMENT FORMATION EN LORRAINE",
+  "ACCUEIL INSERTION FORMATION ORIENTATION",
+  "ACE COMPETENCES",
+  "ACF ALTERNANCE CONSEIL ET FORMATION",
+  "ACFA-MULTIMEDIA",
+  "ACFAL FORMATION",
+  "ACP INTERIM ET RECRUTEMENT",
+  "ACQUIFORMATIONS",
+  "ACT'IN CAMPUS PRO",
+  "ACTION FORMATION",
+  "ACTION FORMATION 81",
+  "ACTIV'CONSULT'TRAINING",
+  "ACTUAL GROUP",
+  "ACTUAL RECRUTE POUR ACTUAL",
+  "ADAFORSS",
+  "ADECCO CENTRE DE FORMATION D'APPRENTIS",
+  "ADEFSA CTRE FORMAT APPRENTIS DESCARTES",
+  "ADEN FORMATIONS ORGANISATION",
+  "ADESA",
+  "ADG",
+  "ADIEL-ACF-ALTERNANCE-CONSEIL-FORMATION",
+  "ADITI",
+  "ADONIS",
+  "ADREC",
+  "AEROCAMPUS AQUITAINE",
+  "AEROKONSULT",
+  "AFA",
+  "AFC  ALTERNANCE FORMATION COMMUNICATION",
+  "AFC GROUPE",
+  "AFEC",
+  "AFI-LNR INDUSTRIES GRAPHIQUES-PAPIER CARTON TRANSFORMATION-PLURIMEDIA",
+  "AFICEPP",
+  "AFIP FORMATIONS",
+  "AFLOKKAT",
+  "AFLOR ACTIMUM",
+  "AFORMANCE",
+  "AFORP FORMATION",
+  "AFPA ENTREPRISES",
+  "AFPI PAYS DE LA LOIRE",
+  "AFPIA-SOLFI2A",
+  "AFRIS PARIS PARMENTIER",
+  "AFTEC",
+  "AFTRAL",
+  "AFUNA",
+  "AG2V MANAGEMENT",
+  "AGENEAU FORMATION",
+  "AGEPAC FORMATION",
+  "AGM FORMATION",
+  "AGORA CONSEIL EN ENTREPRISE",
+  "AGR L'ECOLE DE L'IMAGE",
+  "AIR'LISE FORMATION SAS",
+  "AIROBJECT",
+  "AIVANCITY",
+  "AJ-FORMATION",
+  "AKOR ALTERNANCE",
+  "AKTEAP",
+  "AKXIA",
+  "AL FORMATION",
+  "ALCO FORMATIONS",
+  "ALEO INNOVATION",
+  "ALESIA FORMATION",
+  "ALFAE",
+  "ALINEA FORMATION",
+  "ALKAS FORMATION",
+  "ALL TECHNICS COMMUNICATION",
+  "ALLIANCE CONSEIL PLUS",
+  "ALM SPORT FORMATION",
+  "ALMEA FORMATIONS INTERPRO",
+  "ALMGROUP",
+  "ALORS FORMATION",
+  "ALPES DEVELOPPEMENT FORMATION",
+  "ALPHA COMMUNICATION FORMATION",
+  "ALTER'NATIV",
+  "ALTERNANCE ACADEMY",
+  "ALTERNANCE ALSACE",
+  "ALTERNANCE BASSE NORMANDIE",
+  "ALTERNANCE BOURGOGNE",
+  "ALTERNANCE CORSICA",
+  "ALTERNANCE EXXECC",
+  "ALTERNANCE HAUTES-ALPES",
+  "ALTERNANCE HAUTS DE SEINE",
+  "ALTERNANCE LANGUEDOC",
+  "ALTERNANCE LOIRET",
+  "ALTERNANCE LORRAINE",
+  "ALTERNANCE NORD PAS DE CALAIS",
+  "ALTERNANCE NORMANDIE",
+  "ALTERNANCE PAYS DE LA LOIRE",
+  "ALTERNANCE PICARDIE",
+  "ALTERNANCE PRO",
+  "ALTERNANCE RHONE ALPES",
+  "ALTERNANCE SARTHE",
+  "ALTERNANCE SEINE ET MARNE",
+  "ALTERNANCE SEVRE ET VIENNE",
+  "ALTERNANCE VAR VAUCLUSE",
+  "ALTERNANCE YVELINES",
+  "ALTERNANCIA",
+  "ALTERNATIVE CONSEIL FORMATION",
+  "ALTERNIS BUSINESS SCHOOL",
+  "ALTICOME",
+  "ALXIA",
+  "ALYDRINE",
+  "AMBASSADERM",
+  "AMBRE SELECT ACADEMIE",
+  "AMC",
+  "AMTALENTS",
+  "ANDIL",
+  "ANIM&COM CONSULTANT",
+  "ANTEE FORMATION",
+  "APEPSUP",
+  "APPRENTISSAGE - REUNION - FORMATION. AP-RUN-FORMATION",
+  "APPRENTISSAGE INDUSTRIE AMEUBLEMENT",
+  "APRADIS",
+  "ARCADIA AURA",
+  "ARCADIA BOURGOGNE FRANCHE COMPTE",
+  "ARCHI MED",
+  "ARDEV FORMATION",
+  "ARIANE FORMATION",
+  "ARINFO",
+  "ARINFO I-MAGINER",
+  "ARISLIA",
+  "ARISTEE",
+  "ARTECH FORMATION",
+  "AS FO BEARN SOULE BIGORRE",
+  "ASCENCIA BUSINESS SCHOOL",
+  "ASP BODYGUARD",
+  "ASS  MONTMARTRE NATATION",
+  "ASS ANC ELEVES CONDUCTEURS TRAVAUX",
+  "ASS CMSEA",
+  "ASS DE GESTION DU CENTRE DE FORMATION D'APPRENTIS SUPERIEUR DE BOURGOGNE",
+  "ASS DE GESTION DU CFA REGIONAL DES HAUTS DE FRANCE JEAN BOSCO",
+  "ASS DE L ECOLE DE BIOLOGIE INDUSTRIELLE",
+  "ASS DE PROMOTION ET DE GESTION DES FORMATIONS PROFESSIONNELLES INITIALES EN APPRENTISSAGE POUR LES PME ILE DE FRANCE",
+  "ASS DEP LYCEE PROFESSION PRIVE RURAL AIN",
+  "ASS DES COURS PROF PHARMACIE ACADEMIE",
+  "ASS DES MAISONS FAMILIALES",
+  "ASS DEVELOP APPRENTIS INDUST AQUITAIN",
+  "ASS DROME ARDECHE FORMATION PROFESSION",
+  "ASS ECOLE TECHNIQUE LA CHATAIGNERAIE",
+  "ASS FAMILIALE EDUCATION GATINAIS",
+  "ASS FORMA ACTION SOCIALE ECURIE COURSES",
+  "ASS FORMA ALTERN NEGOCE MAINTENANCE",
+  "ASS FORMA PROF APPRENT AUTO METIERS COMM",
+  "ASS FORMAT AGENTS TOURIS COMMERC INDUST",
+  "ASS FORMAT APPRENTISS SUD-OUEST PARISIEN",
+  "ASS FORMAT DISTRI PRODUIT ELECTRO GD PUB",
+  "ASS FORMAT PROFESS PERMANENTE",
+  "ASS FORMATION APPRENTIS INDUSTRIE",
+  "ASS FORMATION GRANDE DISTRIBUTION",
+  "ASS FORMATION INDUST PAR APPRENTISSAGE",
+  "ASS FORMATION INSERTION GARDOISE",
+  "ASS FORME",
+  "ASS FR DIRECTEURS CHEFS CREDI",
+  "ASS FRANCIL FORMAT PROFES DEVPT APPRENTI",
+  "ASS GEST CTRE FORMATIO APPRENTI PAPETIER",
+  "ASS GEST UNIVERSITE CATHO OUEST BRET SUD",
+  "ASS GESTION CFA CENTRE-ALSACE",
+  "ASS GESTION ECOLE FRANCAISE PAPETERIE",
+  "ASS GESTION INSTITUT FORMATION PROF 43",
+  "ASS INITIA FORMATION APPUI PEDAGOGI EMPL",
+  "ASS IRFA NORMANDIE MAINE",
+  "ASS JEUNES SPORTIFS REGION ILE DE FRANCE",
+  "ASS MAISON FAMIL APPRENTISSAGE",
+  "ASS MAISON FAMILIALE RURALE",
+  "ASS NAT FORMAT PROMOT PROFESS BATELLERIE",
+  "ASS OGEC POITIERS NORD",
+  "ASS OUVRIERE COMPAGNON DEVOIR TOUR FRANC",
+  "ASS POUR DEVELOPPEMENT DE L APPRENTISSAGE DANS LES PROFESSIONS SANITAIRES SOCIALES ET MEDICO SOCIALES EN AQUITAINE",
+  "ASS POUR FORMATION BIOCHIMIE ET BIOLOGIE",
+  "ASS POUR L INSERTION DES JEUNES DANS LE SECTEUR FERROVIAIRE",
+  "ASS POUR LE DEVEL DE L ECHANGEUR REGION",
+  "ASS PROFESSIONS SANITAIRES MEDICO SOCIALES ET DE L INTERVENTION SOCIALE",
+  "ASS PROMO GESTION CTRE ENS SUP A EVRY",
+  "ASS PROMOTION APPRENTISSAGE DANS ENSEIGNEMENT CATHOLIQUE TECHNIQUEAGRICOLE AQUITAINE",
+  "ASS PROMOTION APPRENTISSAGE INDUSTRIES",
+  "ASS REG POUR LA FORMATION D'ANIMATEUR",
+  "ASS ROUEN UNIVERSITE CLUB",
+  "ASSIFEP",
+  "ASSISTANCE SECURITE SYSTEMS",
+  "ASSISTEAL FORMATION",
+  "ASSO FORMAT IMPRIMERIE INDUSTRIE GRAPHIQ",
+  "ASSO MAIS FAM RURALES",
+  "ASSOC DAUPHINOISE FORMATION INDUSTRIE",
+  "ASSOC DU CENTRE DE PROMOTION SOCIALE",
+  "ASSOC INSTITUT MECAN AERAUNOT ET AUTOMOB",
+  "ASSOC INSTITUTION ST ASPAIS DE MELUN",
+  "ASSOC POUR PROMOTION CENTRE TECHNOLOGIE",
+  "ASSOC REG SAUVEGARDE ENFANT ADO ADUL",
+  "ASSOCIATION 1901 FORMATION",
+  "ASSOCIATION DE FORMATION ET D INNOVATION POUR LES METIERS DE L INDUSTRIE ET DE LA MECANIQUE",
+  "ASSOCIATION DE FORMATION POUR LE DEVELOPPEMENT REGIONAL ET L'APPUI AU RECLASSEMENT",
+  "ASSOCIATION DE GESTION DE L' INSTITUT SUPERIEUR D' APPRENTISSAGE DES CCI DE FRANCHE COMTE",
+  "ASSOCIATION DE GESTION DE L'ECOLE SUPERIEURE DES SERVICES",
+  "ASSOCIATION DE GESTION DU CFA DON BOSCO DE NORMANDIE",
+  "ASSOCIATION DE GESTION DU CNAM EN REGION OCCITANIE",
+  "ASSOCIATION DE GESTION DU CONSERVATOIRE NATION DES ARTS ET METIERS AUVERGNE-RHONE-ALPES",
+  "ASSOCIATION DE GESTION DU CONSERVATOIRE NATIONAL DES ARTS ET METIERS DE LA REUNION",
+  "ASSOCIATION DE RECHERCHE ET DE DEVELOPPEMENT POUR LA REUSSITE EDUCATIVE SCOLAIRE ET SOCIALE (A.R.D.R.E.S.S)",
+  "ASSOCIATION DEPARTEMENTALE DES PUPILLES DE L'ENSEIGNEMENT PUBLIC",
+  "ASSOCIATION DES NETICIENS",
+  "ASSOCIATION DON BOSCO LYON",
+  "ASSOCIATION ECOLE SECRETE DE GASTRONOMIE",
+  "ASSOCIATION EDHEC BUSINESS SCHOOL",
+  "ASSOCIATION EDUCASKILLS&FORMASKILLS",
+  "ASSOCIATION EDUCATIVE LA SALLE ST LOUIS STE BARBE",
+  "ASSOCIATION ENFANTS PAYS BELEYME",
+  "ASSOCIATION FORMA ON",
+  "ASSOCIATION GESTIONNAIRE POUR LA FORMATION PROFESSIONNELLE DE LA COIFFURE DE LA HAUTE SAVOIE",
+  "ASSOCIATION GROUPE ESA",
+  "ASSOCIATION IMC ALTERNANCE",
+  "ASSOCIATION ISEP-EDOUARD BRANLY",
+  "ASSOCIATION LES CHENES",
+  "ASSOCIATION O.R.T.",
+  "ASSOCIATION POUR L'EDUCATION COGNITIVE ET LE DEVELOPPEMENT (AECD)",
+  "ASSOCIATION POUR LA FORMATION PROFESSIONNELLE EN MEUSE",
+  "ASSOCIATION POUR LA FORMATION TECHNOLOGIQUE DANS L'ENSEIGNEMENT CATHOLIQUE",
+  "ASSOCIATION POUR LA GESTION DE 3IL",
+  "ASSOCIATION POUR LA MOBILISATION DES SAVOIRS",
+  "ASSOCIATION POUR LA PROMOTION DE L'APPRENTISSAGE DE L'ENSEIGNEMENT CATHOLIQUE TECHNIQUE EN OCCITANIE",
+  "ASSOCIATION POUR LA PROMOTION DE L'APPRENTISSAGE DE L'ENSEIGNEMENT CATHOLIQUE TECHNOLOGIQUE PROFESSIONNEL ET AGRICOLE E",
+  "ASSOCIATION POUR LA PROMOTION DES METIERS DE L'ANIMATION ET DU SPORT GRAND EST",
+  "ASSOCIATION POUR LA PROMOTION SOCIALE ET LA FORMATION PROFESSIONNELLE DANS LES TRANSPORTS ROUTIERS",
+  "ASSOCIATION REGIONALE D'EDUCATION PERMANENTE DU MORBIHAN",
+  "ASSOCIATION REGIONALE DE FORMATION PROFESSIONNELLE DES INDUSTRIES DU BOIS",
+  "ASSOCIATION SAINT JOSEPH",
+  "ASSOCIATION SAINT-YVES / UNIVERSITE CATHOLIQUE DE L'OUEST",
+  "ASSOCIATION SPORT ANIMATION NOUVELLE AQUITAINE",
+  "ASSOCIATION SUD FORMATION",
+  "ASSOCIATION UNIVERSITE ET SPORTS",
+  "ASTERIA",
+  "ASTON INSTITUT",
+  "ATELIER D'AMENAGEMENT TOURISTIQUE POUR L'INSERTION",
+  "ATH FORMATION",
+  "ATHELIA INSTITUT DE FORMATION CONSEILS ET PERFORMANCE",
+  "ATHELIA INSTITUT DE FORMATIONS COMPETENCES ET PERFECTIONNEMENT",
+  "ATHENA BS",
+  "ATK CONSEILS",
+  "ATOUT EMPLOI",
+  "ATOUTS SERVICES ENTREPRISES ET FORMATION",
+  "AUDEO CFA",
+  "AUREIS FORMATION",
+  "AURLOM PREPA",
+  "AUVERGNE FORMATION",
+  "AVENIR FORMATION OCEAN INDIEN",
+  "AVENIR PRO FORMATIONS",
+  "AVLIS FORMATION",
+  "AXIO-LUTION",
+  "B2H 13 AIX",
+  "B2H 83 LA VALETTE",
+  "B2H13 MARSEILLE",
+  "B2H83 FREJUS",
+  "BANLIAT",
+  "BATIMENT CFA BOURGOGNE- FRANCHE- COMTE",
+  "BATIMENT CFA BRETAGNE",
+  "BATIMENT CFA NORMANDIE",
+  "BATISUD FORMATION",
+  "BDL FORMATIONS",
+  "BEAUTY ACADEMIE",
+  "BECKIE'S NAILS",
+  "BEFORMA",
+  "BK AMBITIONS",
+  "BLOOM",
+  "BLOOM ALTERNANCE",
+  "BOULEVARDS DES POTES",
+  "BOURBON FORMATIONS AVENIR",
+  "BP CONSEIL FORMATION",
+  "BRAVUS",
+  "BRUNO TRIAY CONSEIL",
+  "BRUNO TRIAY CONSEIL LENS",
+  "BSA MAUD BADUREAU",
+  "BSF",
+  "BTP CFA AUVERGNE RHONE ALPES",
+  "BTP CFA CENTRE",
+  "BTP CFA GRAND EST",
+  "BTP CFA HAUTS-DE-FRANCE",
+  "BTP CFA ILE DE FRANCE",
+  "BTP CFA LE BELLOY",
+  "BTP CFA NOUVELLE AQUITAINE",
+  "BTP CFA OCCITANIE",
+  "BTP CFA PAYS DE LA LOIRE",
+  "BTP CFA POITOU CHARENTES",
+  "BTP SERVICES",
+  "BUSINESS IMMO",
+  "BZH FORMATION",
+  "C DE FORM APPRENTIS MAISON FAMILIALE",
+  "C E SI",
+  "C. FORMAT PRO CONSEIL",
+  "C.B.M.",
+  "C.F.D.",
+  "C2N CONSEIL ET FORMATION",
+  "C3 CFA",
+  "CA BASSIN D'ARCACHON SUD (COBAS)",
+  "CAB EXPERT",
+  "CAB FORMATIONS",
+  "CABINET ASSISTANCE CONSEILS",
+  "CABINET RAHARISON",
+  "CADIS FORMATIONS",
+  "CAFA FORMATION",
+  "CAMILLI",
+  "CAMPUS ACADEMY ANGERS",
+  "CAMPUS DE L'ESTHETIQUE ET DU SPA",
+  "CAMPUS DU LAC",
+  "CAMPUS FONDERIE DE L'IMAGE",
+  "CAMPUS INTERNATIONAL RIERA",
+  "CAMPUS MECATEAM",
+  "CAMPUS PRO",
+  "CAMPUS PRO CENTRE DE FORMATION DE SAINT- GOBAIN DISTRIBUTION BATIMENT FRANCE",
+  "CAMPUS RIERA",
+  "CAMPUS STRAT@INNOV PARIS",
+  "CAMPUS125",
+  "CAP COMPETENCES - ACTIVATEUR DE TALENTS",
+  "CAP ELAN FORMATION",
+  "CAP FORMATION",
+  "CAP PROFESSIONNALISATION DEPENDANCE ET HANDICAP",
+  "CAPITAL EXPERIENCES",
+  "CAPITAL FORMATION",
+  "CARIBBEAN FOUNDERS INSTITUTE",
+  "CARPROMO",
+  "CARREL CENTRE DE FORMATION",
+  "CARRIERES & TRANSITIONS",
+  "CARRIERES ET CONSEIL",
+  "CATEC CENTRE ARTISTIQUE ET TECHNIQUE DE L'ESTHETIQUE ET DE LA COIFFURE",
+  "CC EXPORT",
+  "CCI FORMATION GERS",
+  "CCS",
+  "CEDEFI-LIDEC L INSTITUT DES ETUDES COMPTABLES",
+  "CEFIM (CENTRE EUROPEEN DE FORMATION INFORMATIQUE  ET MULTIMEDIA)",
+  "CEFIM (CENTRE EUROPEEN DE FORMATION INFORMATIQUE ET MULTIMEDIA)",
+  "CEFODIS AUTO",
+  "CEFOR G",
+  "CEFORA",
+  "CENT EDUCAT PERMANENTE FORMAT CONTINUE",
+  "CENTR FORMATION TECHNICIENS AGRICOLES",
+  "CENTRE   DE FORMATION DES METIERS DES ARTS  DE LA COIFFURE",
+  "CENTRE BIEN ETRE ET BEAUTE INFANTINO",
+  "CENTRE D'ENSEIGNEMENT SUPERIEUR D'ADMINISTRATION ET DE MANAGEMENT DES ENTREPRISES SUP",
+  "CENTRE D'ETUDES ET DE FORMATIONS DES ACTIVITES NAUTIQUES ET TERRESTRES",
+  "CENTRE DE FORMATION ADULTES TX PUBLICS",
+  "CENTRE DE FORMATION APPRENTISSAGE ET PERFECTIONNEMENT",
+  "CENTRE DE FORMATION BEAUSEJOUR",
+  "CENTRE DE FORMATION D'APPRENTIS DE L'INDUSTRIE GRAND OUEST NORMANDIE",
+  "CENTRE DE FORMATION DES APPRENTIS DE L INDUSTRIE LOIRE DROME ARDECHE (CFAI - LDA)",
+  "CENTRE DE FORMATION DES APPRENTIS DE L'EDUCATION POPULAIRE",
+  "CENTRE DE FORMATION DES APPRENTIS DE L'INDUSTRIE FRANCHE-COMTE",
+  "CENTRE DE FORMATION DES METIERS DE L'INDUSTRIE HOTELIERE 54",
+  "CENTRE DE FORMATION DU BORN",
+  "CENTRE DE FORMATION DU GROUPE ACCELIS",
+  "CENTRE DE FORMATION ET D'ENSEIGNEMENT PROFESSIONNEL",
+  "CENTRE DE FORMATION ET D'ORIENTATION PROFESSIONNELLE DE GUYANE",
+  "CENTRE DE FORMATION ET DE PROMOTION",
+  "CENTRE DE FORMATION LAGORD",
+  "CENTRE DE FORMATION LANGUES",
+  "CENTRE DE FORMATION PERMANENTE ST PIERRE",
+  "CENTRE DE FORMATION PRIVE 2S",
+  "CENTRE DE RESSOURCES D'EXPERTISE ET DE PERFORMANCE SPORTIVE",
+  "CENTRE EQUESTRE D'IDOINE (CE D'IDOINE)",
+  "CENTRE FORMAT BOURSE TRAVAIL",
+  "CENTRE HOSPITALIER UNIVERSITAIRE DE BORDEAUX",
+  "CENTRE PEDAGOGIQUE POLYVALENT DE L'UNION",
+  "CENTRE REGIONAL DE FORMATION MULTIPROFESSIONNEL",
+  "CENTRE REUNIONNAIS D'ENTREPRISE EDUCATIVE",
+  "CEPROC",
+  "CERTIFECO",
+  "CESAME",
+  "CESI",
+  "CETALION",
+  "CEVENNES FORMATIONS ALES",
+  "CFA",
+  "CFA DES METIERS DE LA SANTE ET DU SOCIAL AUVERGNE RHONE ALPES",
+  "CFA DU GRAND BERGERACOIS DES METIERS TRADITIONNELS ET DE LA SANTE",
+  "CFA ECOLE SANTE ET JURIDIQUE",
+  "CFA ENGIE",
+  "CFA INSTA",
+  "CFA LUVAL",
+  "CFA MARSEILLE",
+  "CFA PETITE CAMARGUE",
+  "CFA PHARMACIE DE L'AFPPREC - ASSOCIATION POUR LA FORMATION DES PREPARATEURS EN PHARMACIE DE LA REGION CVDL",
+  "CFAI ALSACE",
+  "CFAI D AUVERGNE",
+  "CFAI DE BRETAGNE",
+  "CFAI DE CHAMPAGNE ARDENNE",
+  "CFAI POITOU-CHARENTES",
+  "CFAI REGION NORD PAS-DE-CALAIS",
+  "CFAMC",
+  "CFECC",
+  "CFI FORMATION",
+  "CFI FORMATION - CFI INFORMATIQUE - CFI ENCAISSEMENT",
+  "CFMA",
+  "CFP ECOLE NANTAISE DE COMMERCE-ENC",
+  "CH APPRENTISSAGE METIERS SOIE ET TEXTILE",
+  "CHALON FORMATION",
+  "CHAMB COMMERC INDUSTRIE NICE COTE D'AZUR",
+  "CHAMB COMMERCE ET INDUSTRIE HAUTES ALPES",
+  "CHAMBRE COMMERCE & INDUSTRIE HTE SAVOIE",
+  "CHAMBRE COMMERCE & INDUSTRIE PAU BEARN",
+  "CHAMBRE COMMERCE ET INDUSTRIE GRENOBLE",
+  "CHAMBRE COMMERCE ET INDUSTRIE LA MAYENNE",
+  "CHAMBRE COMMERCE ET INDUSTRIE MORBIHAN",
+  "CHAMBRE COMMERCE INDUSTRIE DE LA DROME",
+  "CHAMBRE DE COMMERCE D INDUSTRIE DE MAINE LOIRE",
+  "CHAMBRE DE COMMERCE ET D INDUSTRIE DE HAUTE LOIRE",
+  "CHAMBRE DE COMMERCE ET D INDUSTRIE DE REGION DES ILES DE GUADELOUPE",
+  "CHAMBRE DE COMMERCE ET D INDUSTRIE DE REGION PARIS ILE DE FRANCE",
+  "CHAMBRE DE COMMERCE ET D INDUSTRIE DES VOSGES",
+  "CHAMBRE DE COMMERCE ET D INDUSTRIE METROPOLE DE BOURGOGNE",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE BAYONNE PAYS BASQUE",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE DE CORSE",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE DE LA REUNION",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE DE REGION HAUTS-DE-FRANCE",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE DES COTES D'ARMOR",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE LE MANS SARTHE",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE METROPOLITAINE BRETAGNE OUEST",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE PUY-DE-DOME CLERMONT AUVERGNE METROPOLE",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE TERRITORIALE ALSACE EUROMETROPOLE",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE TERRITORIALE CHARENTE-MARITIME",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE TERRITORIALE DE L'ALLIER",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE TERRITORIALE DE LA VIENNE",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE TERRITORIALE DU VAR",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE TERRITORIALE ILLE-ET-VILAINE",
+  "CHAMBRE DE COMMERCE ET D'INDUSTRIE TERRITORIALE PORTES DE NORMANDIE",
+  "CHAMBRE DE METIERS ET DE L'ARTISANAT DE REGION BRETAGNE",
+  "CHAMBRE DE METIERS ET DE L'ARTISANAT DE REGION CENTRE-VAL DE LOIRE",
+  "CHAMBRE DE METIERS ET DE L'ARTISANAT DE REGION GRAND EST",
+  "CHAMBRE DE METIERS ET DE L'ARTISANAT DE REGION GUADELOUPE",
+  "CHAMBRE DE METIERS ET DE L'ARTISANAT DE REGION HAUTS-DE-FRANCE",
+  "CHAMBRE DE METIERS ET DE L'ARTISANAT DE REGION ILE-DE-FRANCE",
+  "CHAMBRE DE METIERS ET DE L'ARTISANAT DE REGION NORMANDIE",
+  "CHAMBRE DE METIERS ET DE L'ARTISANAT DE REGION NOUVELLE AQUITAINE",
+  "CHAMBRE DE METIERS ET DE L'ARTISANAT DE REGION OCCITANIE",
+  "CHAMBRE DE METIERS ET DE L'ARTISANAT DE REGION PROVENCE-ALPES-COTE D'AZUR",
+  "CHAMBRE REGIONALE D'AGRICULTURE DES PAYS DE LA LOIRE",
+  "CIFACOM",
+  "CIFCA",
+  "CIFEP LYON",
+  "CIL",
+  "CIPECMA",
+  "CITY FORMATION ACADEMY",
+  "CLASSEOS",
+  "CLM FORMATION",
+  "CLPS L ENJEU COMPETENCES",
+  "CNAM GRAND EST",
+  "CNAM HAUTS-DE-FRANCE - ASSOCIATION DE GESTION",
+  "CNPH PIVERDIERE",
+  "CODEV",
+  "COFAP-CONSEIL FORMATION ASSISTANCE PROFE",
+  "COFORDIS",
+  "COGEFI FORMATION",
+  "COIFFURE CONCEPT",
+  "COLDEFY IPFAC",
+  "COLLEGE DE PARIS GRAND EST",
+  "COLLEGE DE PARIS-ECOLE CONTE",
+  "COMITE LOCAL INSERTION PROFESSIONELLE SOCIALE",
+  "COMMISSION NALE QUALIF ASSIST",
+  "COMMUNAUTE DE COMMUNES DE L'ESTUAIRE",
+  "COMMUNE DE SETE",
+  "COMPETENCES BTP",
+  "COMPETENCES COMMERCE ET INTERNATIONAL",
+  "COMPETENTIS",
+  "CONCEPT FORMATION",
+  "CONCEPT X FORMATION",
+  "CONSEIL FORMATION QUALIFICATION",
+  "CONSERVATOIRE NATIONAL DES ARTS ET METIERS",
+  "CONSERVATOIRE NATIONAL DES ARTS ET METIERS DE LA REGION DES PAYS DE LA LOIRE - ASSOCIATION DE GESTION",
+  "CORTO CONCEPT",
+  "COURS BIOMEDAL",
+  "COURS CAENNAIS PRIVE D'ESTHETIQUE ET DE COIFFURE",
+  "COURS COMMERCIAUX CLOVIS HUGUES",
+  "COURS ET FORMATION EST-REUNION",
+  "COURS SAINT PAUL",
+  "COURS SUPERIEUR DE COIFFURE",
+  "COUSSIERE",
+  "CPA LATHUS",
+  "CPES - CLASSES PREPARATOIRES AUX ETUDES DE SANTE",
+  "CPO.A BTP",
+  "CREAFI",
+  "CREALISE",
+  "CREATIVE",
+  "CROIX ROUGE FRANCAISE",
+  "CRS PROF PERFECTIONNEMENT PHARMACIE",
+  "CTI",
+  "CTR FORMATION APPRENTI INTERPROF COROT",
+  "CTRE CARAIBEEN DEVELOPPEMENT COMPETENCES",
+  "CTRE CONSULAIRE FORMATION PROFESSION AIN",
+  "CTRE EDUCATION ROUTIERE FORMAT CONTINUE",
+  "CTRE EUROPEEN FORMAT CONTINUE MARITIME",
+  "CTRE EUROPEEN FORMATION PROFESS HOTELIER",
+  "CTRE EUROPEEN RECHER FORMA ARTS DU VERRE",
+  "CTRE FORMAT APPRENT INDUST MIDI PYRENEES",
+  "CTRE RECHERCHES ETUDES FORMATION ORGANIS",
+  "CTRE SUPERIEUR FORMAT APPRENTISSAGE.BTP",
+  "CURSANOVA",
+  "CVEDUCATION",
+  "D2J",
+  "DATASCIENTEST",
+  "DAUDE",
+  "DAWAN",
+  "DBD FORMATION & CONSULTING",
+  "DCF",
+  "DEA FORMATION",
+  "DELTA",
+  "DEVANA",
+  "DEVELOP'A OU",
+  "DEVELOPPEMENT FORMATION PROXIMITE 971",
+  "DGBOOST",
+  "DIAPASONIA",
+  "DIDEROT EDUCATION CAMPUS",
+  "DIFCAM",
+  "DIGINAMIC",
+  "DIGITAL LEARNING CONTEST",
+  "DIJON FORMATION",
+  "DIPLOMA SUP",
+  "DIRECTT FORMATION",
+  "DISTRICOM FORMATION",
+  "DL-FORMATION",
+  "DOCEO FORMATION",
+  "DOMINO RH CARE VERSAILLES",
+  "DORANCO ESPACE MULTIMEDIA",
+  "DRAC",
+  "DREAM ACADEMY",
+  "DUCOS FABIEN FORMATION",
+  "E S G M ECOLE SUP GESTION MANAGEMENT",
+  "E-ART SUP INSTITUT",
+  "E-BUSINESS SCHOOL",
+  "E-CATALYST",
+  "E-WORKS",
+  "E2M FORMATION",
+  "E2SE BUSINESS SCHOOL",
+  "EA PLUS FORMATION",
+  "EBBS-BUSINESS SCHOOL",
+  "ECEC",
+  "ECEMA VALENCE",
+  "ECF CENTRE OUEST ATLANTIQUE",
+  "ECM",
+  "ECOFAC EVOLUTION",
+  "ECOLE ACTIVE ET INTERNATIONALE",
+  "ECOLE AGRICOLE PRIVEE MAISON FAMI RURALE",
+  "ECOLE BOULANGERIE ET PATISSERIE DE PARIS",
+  "ECOLE CENTRALE DE LYON",
+  "ECOLE CENTRALE DE NANTES",
+  "ECOLE D ENSEIGNEMENT SUPERIEUR CONSULAIRE DE MEURTHE ET MOSELLE",
+  "ECOLE D INGENIEURS DE PURPAN",
+  "ECOLE D'ARTS APPLIQUES",
+  "ECOLE D'ENSEIGNEMENT SUPERIEUR PRIVE DES ARTS DE LA COSMETOLOGIE ET DE L'ESTHETIQUE BEAUTE THALGO INTERNATIONALE",
+  "ECOLE DE COMMERCE ET DE MANAGEMENT DE MULHOUSE",
+  "ECOLE DE CONDUITE FRANCAISE COOPERATIVE D'EDUCATION ROUTIERE CENTRE ATLANTIQUE- E.C.F. CERCA",
+  "ECOLE DE LA CONSTRUCTION DES INFRASTRUCTURES ET RESEAUX APPRENTISSAGE",
+  "ECOLE DE LA CONSTRUCTION DES INFRASTRUCTURES ET RESEAUX FORMATION",
+  "ECOLE DE MANAGEMENT DE CRETEIL",
+  "ECOLE DENTAIRE FRANCAISE",
+  "ECOLE DES COMPETENCES",
+  "ECOLE DES FORMATIONS MARITIMES",
+  "ECOLE DES HAUTES ETUDES EN SANTE PUBLIQUE",
+  "ECOLE DES METIERS DE L'ALIMENTATION ET DU COMMERCE",
+  "ECOLE DES METIERS DU SERVICE  DU COMMERCE ET DE L'ACCUEIL",
+  "ECOLE DES METIERS GBH",
+  "ECOLE ENSEIGNEMENT SUPERIEUR PRIVE ICN",
+  "ECOLE FINISTERIENNE COMMERCE GESTION",
+  "ECOLE FORMATION INITIALE INDUS PHARMACEU",
+  "ECOLE FR HOTESSE ET DE TOURISME - EFHT",
+  "ECOLE FRANCAISE DE L'ALTERNANCE",
+  "ECOLE GUSTAVE",
+  "ECOLE INTERNATIONALE DE MANAGEMENT DE PARIS EIMP",
+  "ECOLE INTERNATIONALE DU SAVOIR FAIRE FRANCAIS",
+  "ECOLE MORBIHANNAISE DE COMMERCE ET GESTION",
+  "ECOLE NAT SUPER CREATION INDUSTRIELLE",
+  "ECOLE NATIONALE DE SECURITE - NIMES",
+  "ECOLE NATIONALE SUPERIEURE D'ARCHITECTURE DE PARIS-EST",
+  "ECOLE NATIONALE SUPERIEURE D'ARCHITECTURE ET DE PAYSAGE DE BORDEAUX",
+  "ECOLE NATIONALE SUPERIEURE DE PAYSAGE DE VERSAILLES",
+  "ECOLE NATIONALE SUPERIEURE DES METIERS DE LA VIANDE",
+  "ECOLE NORMALE SUPERIEURE PARIS-SACLAY",
+  "ECOLE PRIVEE D' ESTHETIQUE DE TOURAINE",
+  "ECOLE PRIVEE DE COIFFURE DE LYON",
+  "ECOLE PRIVEE ET CENTRE DE FORMATION D APPRENTIS ESTHETIQUE ET BIEN - ETRE CAMILLE SELINE",
+  "ECOLE PROFESSIONNELLE DE BOUCHERIE",
+  "ECOLE REGIONALE DE LA FORMATION PROFESSIONNELLE",
+  "ECOLE SCIENCES & TECHNIQUES COMMERCIALES",
+  "ECOLE SUP INFORM  COMMERCE",
+  "ECOLE SUPERIEUR COMPTABILITE GESTION",
+  "ECOLE SUPERIEUR DE COMMERCE FLORANGE",
+  "ECOLE SUPERIEUR DES SERVICES",
+  "ECOLE SUPERIEURE COMMERCE COMMUNICATION",
+  "ECOLE SUPERIEURE D'ECONOMIE RESPONSABLE",
+  "ECOLE SUPERIEURE D'INFORMATIQUE",
+  "ECOLE SUPERIEURE D'INFORMATIQUE - SUPINFO PARIS",
+  "ECOLE SUPERIEURE DE COIFFURE",
+  "ECOLE SUPERIEURE DE COMMERCE DES AFFAIRES ET DU MANAGEMENT",
+  "ECOLE SUPERIEURE DE COMMERCE DES TROIS FRONTIERES",
+  "ECOLE SUPERIEURE DE COMMERCE ET DE COMMUNICATION",
+  "ECOLE SUPERIEURE DE COMMERCE ET GESTION-ESCG PARIS",
+  "ECOLE SUPERIEURE DE LA BIO",
+  "ECOLE SUPERIEURE DE MANAGEMENT DES LANDES",
+  "ECOLE SUPERIEURE DE MANAGEMENT ET DE COMPTABILITE",
+  "ECOLE SUPERIEURE DES COMPETENCES ET METIERS",
+  "ECOLE SUPERIEURE DES METIERS DU TERTIAIRE",
+  "ECOLE SUPERIEURE DES SERVICES AU ENTREPRISES",
+  "ECOLE SUPERIEURE DES TECHNOLOGIES INDUSTRIELLES AVANCEES - ESTIA",
+  "ECOLE SUPERIEURE DU FOOD MANAGEMENT",
+  "ECOLE TECHNIQUE PRIVEE COIFFURE JORLIS",
+  "ECOLE TECHNIQUE PRIVEE EUROPEENNE D'ESTHETIQUE",
+  "ECOLE TECHNIQUE SUPER CHIMIE DE L'OUEST",
+  "ECOLE TOURANGELLE SUPERIEURE ETABLISSEMENT PRIVE E.T.S.",
+  "ECOLE VIDAL SARL",
+  "ECOLE WEB ET MOBILE",
+  "ECOLES BILLIERES",
+  "ECOR / ECAR",
+  "ECORIS",
+  "ECOTEC",
+  "ECP APPRENTISSAGE",
+  "ECPM",
+  "EDA GROUPE",
+  "EDACADEMY",
+  "EDAIC",
+  "EDUC'ACTION",
+  "EDUC'ALTITUDES",
+  "EDUSKALE",
+  "EFFOR ACADEMY",
+  "EFFOR SPORT & BUSINESS",
+  "EFIA",
+  "EFREI PARIS",
+  "EHO L'ECOLE HOTELIERE",
+  "EI GROUPE",
+  "ELAN FORMATION",
+  "ELEGANCE",
+  "ELOCE",
+  "ELYSEES CANNES",
+  "EMG BUSINESS SCHOOL",
+  "EMGF",
+  "EMIC FORMATION",
+  "EMPORUS",
+  "EMSAT (ECOLE DES METIERS DU SPORT, DE L'ANIMATION ET DU TOURISME)",
+  "ENCP",
+  "ENERGO GENIUS",
+  "ENERGYZON",
+  "ENGAGEMENT & ENTREPRISE",
+  "ENS NOTRE DAME SAINT JOSEPH",
+  "ENSEIGNEMENT FORMATION INSERTION PROMO",
+  "ENSUP",
+  "EPA CITE DE LA FORMATION MARMANDE",
+  "EPEC GASPARD",
+  "EPEC ST ETIENNE",
+  "EPINAY MANAGEMENT SCHOOL",
+  "EPITECH ECOLE INFORMATIQUE NOUV TECHNOL",
+  "EPLEFPA DE CROIX-RIVAIL",
+  "EQUIS",
+  "ERENA ACADEMIA",
+  "ERICS ASSOCIES",
+  "ES. ANTOINE GAPP",
+  "ESAGE",
+  "ESC'AL FORMATION",
+  "ESCALE",
+  "ESCG MARSEILLE",
+  "ESCG MONTPELLIER",
+  "ESED",
+  "ESGCV",
+  "ESID",
+  "ESIMA LYON - PERFORMANCES CONCEPT",
+  "ESIMED",
+  "ESMED - AFTEC SORGUES",
+  "ESPACE REUSSITE",
+  "ESPIC",
+  "ESPL-ECOLE SUPERIEURE DES PAYS DE LOIRE",
+  "ESS-RH",
+  "ESSOR",
+  "ESTHETIC AND COMPANY",
+  "ESTHETIQUE FORMATION SARL",
+  "ESUP GROUP",
+  "ETAB PUBLIC ENSEIGNEMENT AGRICOLE OISE",
+  "ETABLISSEMENT D'ENSEIGNEMENT SUPERIEUR DE GESTION ET DE L'ENVIRONNEMENT",
+  "ETABLISSEMENT PUBLIC LOCAL D ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLE DE PAU MONTARDON",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION AGRICOLE MIRANDE-RISCLE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLE D'ANTIBES",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLE DE CASTELNAUDARY",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLE DE CHARTRES - LA SAUSSAYE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLE DE CHATEAUROUX",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLE DE L'OISELLERIE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLE DE LA MEUSE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLE DE MEURTHE ET MOSELLE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLE DE MONTLUCON LAREQUILLE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLE DE PONTIVY-ST JEAN BREVELAY-HENNEBONT",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLE DE VALDOIE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES CHAMBERY LA MOTTE-SERVOLEX (SAVOIE)",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES CHATEAULIN MORLAIX KERLIVER",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DE BORDEAUX GIRONDE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DE CIBEINS",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DE L'EURE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DE LA ROCHE SUR YON",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DE LONS-LE-SAUNIER",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DE LYON-DARDILLY - RHONE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DE MERDRIGNAC",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DE MONTPELLIER-ORB-HERAULT",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DE POITIERS-VENOURS",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DE RENNES - LE RHEU",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DE SAINT-AFFRIQUE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DU LOIR ET CHER",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES DU VELAY",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET DE FORMATION PROFESSIONNELLE AGRICOLES HAUTE CORREZE",
+  "ETABLISSEMENT PUBLIC LOCAL D'ENSEIGNEMENT ET FORMATION PROFESSIONNELLE AGRICOLES D'AUCH BEAULIEU LAVACANT",
+  "ETNA ECOLE TECH NUMERIQUES APPLIQUES",
+  "ETP DUCLAUX",
+  "ETS PUBLIC NATIONAL A KOENIGSWARTER",
+  "EUCLEA",
+  "EURIDIS MANAGEMENT",
+  "EURO MEDIA FORMATION",
+  "EURO TEAM CAPELLE",
+  "EUROCOM STRATEGIES",
+  "EUROFORMATION - CONSULTANCE",
+  "EUROFORMATION SUD",
+  "EUROPEAN CONSULTING MANAGEMENT",
+  "EVEREN FORMATION",
+  "EVOCIME FORMATIONS COMPORTEMENTALES",
+  "EVOLUA FORMATION",
+  "EVOLUTION FORMATION OCEAN INDIEN",
+  "EVOYA CONSEILS",
+  "EXACTETUDES",
+  "EXCEL FORMATION - L'ECOLE DU COMMERCE ET DU SERVICE",
+  "EXCELLENCE ACADEMY",
+  "EXCELLENCE BS",
+  "EXCELLENTIA FORMATION & ASSOCIEES",
+  "EXPERNET",
+  "EXPERSONA",
+  "EXPERT BUSINESS FORMATIONS",
+  "EXPERT STRATEGIE ACADEMY",
+  "EXXEA",
+  "FACULTE DES METIERS DE L ESSONNE",
+  "FANFANI",
+  "FB FORMATION",
+  "FBS - FMS",
+  "FC2R",
+  "FCF-ARGOS",
+  "FCIP EDUCATION ET FORMATION TOUT AU LONG DE LA VIE",
+  "FEDER COMPAGNON METIERS BATIMENT",
+  "FEDER COMPAGNONNIQUE BATIMENT",
+  "FEDERATION COMPAGNONNIQUE REGIONALE OCCITANIE",
+  "FEDERATION DES ARTISANS BOULANGERS PATISSIERS DU PUY DE DOME",
+  "FEDERATION FRANCAISE D EQUITATION",
+  "FEDERATION FRANCAISE DE NATATION",
+  "FENELON SAINTE-MARIE-LA PLAINE MONCEAU",
+  "FERGUSS ACADEMIE",
+  "FIRST AVENIR",
+  "FONDATION APPRENTIS D'AUTEUIL",
+  "FONDATION EDITH SELTZER",
+  "FOOD'N CO LEARNING",
+  "FOR YOU ONLY",
+  "FOR'MISSION",
+  "FORCE FORMAT COMPET EFFIC SAVOIR FAIRE",
+  "FORE ILES DU NORD",
+  "FOREACH ACADEMY",
+  "FOREZ INSERTION FORMATION CONSEIL",
+  "FORM C' PRO FORMATION CONTINUE PROF",
+  "FORM HIGH TECH",
+  "FORM' ET VOUS",
+  "FORM'A OU",
+  "FORM'AQUI",
+  "FORM'INTUITIVE",
+  "FORM.ECARE",
+  "FORMA CONSEIL",
+  "FORMA DRINKS",
+  "FORMA OUEST INSERTION",
+  "FORMA PLUS",
+  "FORMA SANTE",
+  "FORMA SMART",
+  "FORMA' INSTITUT DE FORMATION DE LA FSCF",
+  "FORMAGRAPH DESIGN",
+  "FORMALIFE ACCOMPAGNEMENT PROFESSIONNEL ET FORMATION",
+  "FORMAMOD",
+  "FORMAOUEST",
+  "FORMAPOSTE MIDI ATLANTIQUE",
+  "FORMARENA",
+  "FORMASEQ",
+  "FORMASO",
+  "FORMASPORT VAR",
+  "FORMASUP 82",
+  "FORMASUP AUVERGNE",
+  "FORMASUP HAUTS DE FRANCE",
+  "FORMASUP ISERE DROME ARDECHE",
+  "FORMASUP PROVENCE ALPES COTE D'AZUR",
+  "FORMASUP TOURISME",
+  "FORMAT DIFFERENCE",
+  "FORMATAGE LANGUEDOC",
+  "FORMATION & CONSEIL",
+  "FORMATION ALTERNANCE COMMERCE",
+  "FORMATION COMPAGNONNIQUE DES PAYS DE LA LOIRE",
+  "FORMATION COMPAGNONNIQUE METIER BATIMENT",
+  "FORMATION CREATIONS",
+  "FORMATION DEMOCRATIE NORMANDIE HAVRE",
+  "FORMATION ET METIER",
+  "FORMATION INNOVANTE EDUCATIVE RESPONSABLE",
+  "FORMATION OI HOTELLERIE RESTAURATION BAR",
+  "FORMATION ORGANIS RECRUTEMENT ACTION",
+  "FORMATION PAR EXCELLENCE ALTERNEE",
+  "FORMATION PLUS MEDITERRANEE",
+  "FORMATION PROFES OUVRIER MAREE",
+  "FORMATIONS AUX METIERS D'AVENIR",
+  "FORMATIONS D'AKI10",
+  "FORMATIVES",
+  "FORMAVAR",
+  "FORMAVENIR ET RECRUTEMENT",
+  "FORMAZOT OCEAN INDIEN",
+  "FORMETIK",
+  "FPA",
+  "FPA-HSE",
+  "FRANCE LANGUES COMMUNICATION",
+  "FRANCE ORIENTATION CONSEIL",
+  "FRIED CHICKEN RUN FORMATION",
+  "FUTURA",
+  "G&L FORMATION",
+  "G2R",
+  "GALAXIE 5",
+  "GALILEO GLOBAL EDUCATION FRANCE DEVELOPPEMENT 1",
+  "GARTI",
+  "GEFI",
+  "GENESIS PROGRESSIVE ELABORATION DE COMPETENCES",
+  "GEST ECOL FORM INITIAL ALTER METIE TP",
+  "GESTIC FORMATION",
+  "GESTION FORMATION CONSEIL-AUVERGNE",
+  "GFA-REUNION",
+  "GIE CFA DES CHEFS",
+  "GIP - FORMATION CONTINUE ET INSERTION PROFESSIONNELLE DE L'ACADEMIE DE POITIERS",
+  "GIP AUVERGNE",
+  "GIP CAMPUS ESPRIT INDUSTRIES",
+  "GIP FORINVAL DE L'ACADEMIE D'AMIENS",
+  "GIP FORMATION CONTINUE ET INSERTION PROF",
+  "GIP FORMATION CONTINUE ET INSERTION PROFESSIONNELLE DE L'ADADEMIE DE MARTINIQUE",
+  "GIP FORMATION CONTINUE ET INSERTION PROFESSIONNELLE REUNION",
+  "GIP FORMATION TOUT AU LONG DE LA VIE",
+  "GIP FORMATION TOUT AU LONG DE LA VIE ET INSERTION PROFESSIONNELLE DE L ACADEMIE D ORLEANS TOURS",
+  "GIP FORMAVIE",
+  "GLOBAL SPORT BUSINESS",
+  "GLOBAL TRAINING FORMATION",
+  "GLOBE 93 FORMATION",
+  "GMD",
+  "GMS FORMATION",
+  "GOLDENCOLLAR",
+  "GREEN SPARK FORMATION",
+  "GROUPE ALTERNANCE",
+  "GROUPE ALTERNANCE NICE",
+  "GROUPE ASSIFEP",
+  "GROUPE COACHING ESTHETIQUE FRANCE",
+  "GROUPE CONSEIL ASSURANCES FORMATION",
+  "GROUPE CONSEIL INGENIERIE FORMATION",
+  "GROUPE DE LA SALLE - REIMS",
+  "GROUPE ECOLE SUPERIEURE DES PROFESSIONS IMMOBILIERES",
+  "GROUPE ESP-ESD",
+  "GROUPE GEMA - ESI BUSINESS SCHOOL / IA SCHOOL",
+  "GROUPE IGF",
+  "GROUPE IGS - ADIP",
+  "GROUPE KEDGE BUSINESS SCHOOL",
+  "GROUPE MAB",
+  "GROUPE MMR EDUCATION",
+  "GROUPE RENNES SCHOOL OF BUSINESS",
+  "GROUPE SCOLAIRE D'ENSEIGNEMENT AGRICOLE  PRIVE ANTOINE DE SAINT-EXUPERY",
+  "GROUPE SOMEFORM",
+  "GROUPE SUP DE CO AMIENS PICARDIE",
+  "GROUPEMENT D INTERET ASSOCIATIF POUR L APPRENTISSAGE EN TRAVAIL SOCIAL",
+  "GROUPEMENT INTERPROFESSIONNEL POUR L'APPRENTISSAGE ET LA FORMATION CONTINUE",
+  "GROUPEMENT NATIONAL FORMATION AUTOMOBILE",
+  "GROUPEMENT PROMOTION ECHANGES INTERNATIONAUX",
+  "GSM MASTER",
+  "H & C CONSEIL",
+  "H ACADEMIE",
+  "H3",
+  "HABITAT FORMATION",
+  "HAUTE ECOLE DU TRAVAIL ET DE L'INTERVENTION SOCIALE",
+  "HAUTES ETUDES APPLIQUEES DU DROIT (HEAD)",
+  "HBE",
+  "HECTEN",
+  "HELP",
+  "HENRIMAN FORMATION",
+  "HETIC",
+  "HEXAGONE",
+  "HILOZA",
+  "HISSEZ-HAUT",
+  "HNCCBS",
+  "HOMMES ET SAVOIRS",
+  "HP FORMATION",
+  "HR CONSEIL",
+  "HUBACADEMY",
+  "HUI EQUISERVICES",
+  "HUPSO",
+  "HYGIE FORMATIONS PHARMACIE D AQUITAINE CFA ET FORMATION CONTINUE",
+  "I LEV OPEN CAMPUS",
+  'I P L "ISEFAC PARIS LILLE"',
+  "I.C.E.E. INSTITUT DES CARRIERES EUROPEENNES DE L'EXPERTISE",
+  "I.E.S.A.",
+  "I.F.A.P.E. INSTITUT DE FORMATION ET D'ACTION PROFESSIONNELLE EUROPEEN",
+  "I.F.C. DIS GROUPE",
+  "I.F.P. ATLANTIQUE",
+  "I.F.T.E AIX.",
+  "I.N.P.E.C",
+  "IAFM/PACKWELL FORMATION",
+  "ICADEMIE",
+  "ICAM OUEST",
+  "ICCI FORMATIONS",
+  "ICEP",
+  "ICF ATLANTIQUE",
+  "ICF ATLANTIQUE FORMATION ST NAZAIRE",
+  "IDEA IMMO",
+  "IDEKO FORMATION",
+  "IEF2I",
+  "IF PRO",
+  "IF2M",
+  "IFAP MANAGEMENT",
+  "IFCE INSTITUT FORMATION COMPTABLE EUROPE",
+  "IFCV",
+  "IFCV APPRENTISSAGE",
+  "IFDA",
+  "IFI PEINTURE",
+  "IFIR ARL (AIN-RHONE-LOIRE)",
+  "IFOLOG MEDITERRANEE",
+  "IFORM",
+  "IFORMA SUD",
+  "IFP ENERGIES NOUVELLES",
+  "IFP'ALTERNANCE",
+  "IFR@SS",
+  "IFRES (INSTITUT DE FORMATION ROCHES EN SIDOBRE)",
+  "IFRIA BOURGOGNE-FRANCHE COMTE",
+  "IFRIA NORD PAS DE CALAIS - PICARDIE",
+  "IFTE MARSEILLE",
+  "IIT FORMATIONS ALTERNEES ORNES",
+  "IMC RANDSTAD",
+  "IMDTEC",
+  "IMIE PARIS",
+  "IN'FORMA",
+  "INALTA FORMATION",
+  "ING CONSEIL ET FORMATIONS",
+  "INGENIORS",
+  "INGETIS",
+  "INITIATIVES",
+  "INITIATIVES FORMATION",
+  "INNOV'EDUC",
+  "INNOVAFORMATION",
+  "INNOVATION DEVELOPPEMENT FORMATION",
+  "INSIDE CAMPUS",
+  "INST DE FORMATION COMMERCIALE PERMANENTE",
+  "INST FORM REG INDUST AGRO ALIMENT AUVERGNE RHONE ALPES",
+  "INST FORMA REGION INDUS AGRO ALIMEN PACA",
+  "INST INTERC FORMAT ALTER METIERS AUTOMOB",
+  "INST NAT FORMAT APPLIC CTRE CULTURE OUVR",
+  "INST.DEVELOP. COMPETENCES PROF.IDC PRO",
+  "INSTA",
+  "INSTI FORMA SUPERIEURE CEVENNES",
+  "INSTIC / 3IS LYON",
+  "INSTITUT AERO FORMATIONS",
+  "INSTITUT BRETON D'EDUCATION PERMANENTE",
+  "INSTITUT CATHOLIQUE",
+  "INSTITUT CATHOLIQUE DE FORMATION PAR ALTERNANCE",
+  "INSTITUT COOPERATIF DE L APPRENTISSAGE DE LA RECHERCHE ET DE L ENSEIGNEMENT",
+  "INSTITUT D ENSEIGNEMENT SUPERIEUR TECHNIQUE",
+  "INSTITUT D'ADMINISTRATION DES ENTREPRISES DE PARIS",
+  "INSTITUT DE FORMATION AVANCEE AUX METIERS DE LA SANTE",
+  "INSTITUT DE FORMATION DE LA REUNION",
+  "INSTITUT DE FORMATION DES ETABLISSEMENTS TOURISTIQUES ET HOTELIERS",
+  "INSTITUT DE FORMATION DES METIERS DE LA PHARMACIE",
+  "INSTITUT DE FORMATION EN ALTERNANCE DE BOURGES",
+  "INSTITUT DE FORMATION EN COMMERCE & OENOLOGIE",
+  "INSTITUT DE FORMATION ET DE CONSEIL EN GESTION DE CARRIERES",
+  "INSTITUT DE FORMATION PRATIQUE",
+  "INSTITUT DE FORMATION PROFESSIONNELLE EN REGION ACADEMIQUE",
+  "INSTITUT DE FORMATION REGIONAL DES INDUSTRIES ALIMENTAIRES DE L'OUEST",
+  "INSTITUT DE FORMATION REGIONAL DES INDUSTRIES ALIMENTAIRES DE NOUVELLE AQUITAINE",
+  "INSTITUT DE FORMATION REGIONAL DES INDUSTRIES ALIMENTAIRES OCCITANIE",
+  "INSTITUT DE FORMATION SUPERIEUR DE L'ENSEIGNEMENT CONTINUE OCEAN INDIEN",
+  "INSTITUT DE GENECH",
+  "INSTITUT DE GESTION SOCIALE",
+  "INSTITUT DE L'IMAGE DE L'OCEAN INDIEN",
+  "INSTITUT DE LA FONCTION COMMERCIALE",
+  "INSTITUT DE LA RELATION CLIENT - SUD",
+  "INSTITUT DE MANAGEMENT D'AUBERVILLIERS",
+  "INSTITUT DE SOUDURE INDUSTRIE",
+  "INSTITUT DES COMPETENCES ET DES METIERS DE DEMAIN",
+  "INSTITUT DES EDUCATEURS",
+  "INSTITUT DES METIERS DE L'ENSEIGNEMENT SUPERIEUR",
+  "INSTITUT DES METIERS DE LA FORME",
+  "INSTITUT DES METIERS NETWORK",
+  "INSTITUT DES TECHNOLOGIES INFORMATIQUES ET DES SERVICES AUX ENTREPRISES",
+  "INSTITUT DU MARAIS - CHARLEMAGNE - POLLES",
+  "INSTITUT DU TRAVAIL SOCIAL PIERRE BOURDIEU",
+  "INSTITUT EUROPEEN F 2I",
+  "INSTITUT F2I",
+  "INSTITUT FORMAT REGION INDUS AGROALIMENT",
+  "INSTITUT FORMATION AFFAIRES ET GESTION",
+  "INSTITUT FORMATION ASSISTANCE PERSONNES",
+  "INSTITUT FORMATION PROFESSION PERMANENTE",
+  "INSTITUT FORMATION RHONE ALPES",
+  "INSTITUT FORMATION SANTE DE L OUEST",
+  "INSTITUT FRANCAIS DE FINANCE ET DE DATA",
+  "INSTITUT FRANCAIS DE FORMATION PROFESSIONNELLE A&A",
+  "INSTITUT FRANCAIS DES AFFAIRES",
+  "INSTITUT INTERNATIONAL DES NOUVELLES TECHNOLOGIES",
+  "INSTITUT KEOLIS",
+  "INSTITUT LEGES",
+  "INSTITUT MANAGERIAL SUPERIEUR DE COMMERCE ET DE FORMATION",
+  "INSTITUT MARIE THERESE SOLACROUP",
+  "INSTITUT MINES TELECOM",
+  "INSTITUT MINES-TELECOM",
+  "INSTITUT NATIONAL D'ENSEIGNEMENT SUPERIEUR POUR L'AGRICULTURE L'ALIMENTATION ET L'ENVIRONNEMENT",
+  "INSTITUT NATIONAL DE FORMATION ET D'APPLICATION DU CENTRE DE CULTURE OUVRIERE - INFA",
+  "INSTITUT NATIONAL DE L'HYGIENE ET DU NETTOYAGE INDUSTRIEL",
+  "INSTITUT NATIONAL DE L'INFORMATION GEOGRAPHIQUE ET FORESTIERE",
+  "INSTITUT NATIONAL DES ARTS DU MUSIC HALL",
+  "INSTITUT NATIONAL DES FORMATIONS NOTARIALES",
+  "INSTITUT NATIONAL DES SCIENCES APPLIQUEES CENTRE VAL DE LOIRE",
+  "INSTITUT NATIONAL DES SCIENCES APPLIQUEES DE STRASBOURG",
+  "INSTITUT NATIONAL DES SCIENCES APPLIQUEES DE TOULOUSE",
+  "INSTITUT NATIONAL DU CYCLE ET MOTOCYCLE",
+  "INSTITUT NATIONAL UNIVERSITAIRE JEAN-FRANCOIS CHAMPOLLION",
+  "INSTITUT NAUTIQUE BRETAGNE",
+  "INSTITUT PARISIEN DE FORMATION",
+  "INSTITUT POLYTECHNIQUE UNILASALLE",
+  "INSTITUT PRIVE D'ALTERNANCE ET DE MANAGEMENT",
+  "INSTITUT REG UNIVERSITAIRE POLYTECHNIQUE",
+  "INSTITUT REGIONAL D'ETUDES DE TECHNIQUES APPIQUEES",
+  "INSTITUT REGIONAL DE FORMATION DU COMITE REGIONAL DE NORMANDIE DE GYMNASTIQUE",
+  "INSTITUT REGIONAL POUR LA FORMATION D ADULTES DE L EST",
+  "INSTITUT RESSOURCES INTERVENTION SOCIALE",
+  "INSTITUT SUP COMMERCE INFORMATIQUE ORSAY",
+  "INSTITUT SUPERIEUR COMMERCE GESTION",
+  "INSTITUT SUPERIEUR D ETUDES FINANCIERES ET DE MANAGEMENT APPLIQUE",
+  "INSTITUT SUPERIEUR D'INFORMATIQUE ET DE MANAGEMENT DE L'INFORMATION",
+  "INSTITUT SUPERIEUR D'OPTIQUE DE NANCY",
+  "INSTITUT SUPERIEUR DE FORMATION ET DE PREPARATION-ISFP",
+  "INSTITUT SUPERIEUR DE GESTION",
+  "INSTITUT SUPERIEUR DE GESTION COMMERCE ET NUMERIQUE",
+  "INSTITUT SUPERIEUR DE TECHNIQUES ECONOMIQUE COMPTABLE GESTION ET RECHERCHE APPLIQUEE DE BRETAGNE (ISTECGRA BRETAGNE",
+  "INSTITUT SUPERIEUR DES TECHNIQUES DE LA PERFORMANCE",
+  "INSTITUT SUPERIEUR SPECIAL DE LA MODE",
+  "INSTITUT SUPERIEUR TECHNIQUE PRIVE...",
+  "INSTITUT TECHNOLOGIES ET COMMUNICATION",
+  "INSTITUT UNIVERSITAIRE TECHNICIEN SUPERI",
+  "INSY2S",
+  "INTEGRALE SECURITE FORMATIONS",
+  "INTERACTIVE BUSINESS SCHOOL ALTERNANCE",
+  "INTERI",
+  "INTERNATIONAL BUSINESS SOFTWARE GLOBAL SERVICES",
+  "INTERNATIONAL TERRA INSTITUTE",
+  "IPAC",
+  "IPAG",
+  "IPF CONSEIL ET RESSOURCES",
+  "IPSO FORMATIONS",
+  "IQC SECURITY CONSULTANCY",
+  "IRFA SUD",
+  "IRON BODYFIT ACADEMY",
+  "IRSS",
+  "IRSS SANTE - PROFILSUP",
+  "IRSS SPORT",
+  "ISAL",
+  "IMMERSIA",
+  "ISCOD",
+  "ISCOD ALTERNANCE",
+  "ISEAH FORMATION",
+  "ISEC FORMATION",
+  "ISEQ",
+  "ISFAC",
+  "ISIM",
+  "ISIM 83",
+  "ISPN",
+  "ISTEA CONSEIL",
+  "ISTEG",
+  "ISTF",
+  "ISUPCOMPTA",
+  "ISW EDUCATION",
+  "ITIC PARIS",
+  "J'AI VOTRE SOLUTION",
+  "JARDY EQUITATION",
+  "JBB",
+  "JBM",
+  "JFA & ASSOCIES",
+  "JMJ FORMATION",
+  "JOUR J RECRUTEMENT",
+  "JRT FORMATION",
+  "JT CONSEIL FORMATION",
+  "JULIEN CORPORATE ACADEMIE",
+  "JURISCAMPUS",
+  "KAISCHOOL",
+  "KAPTIV",
+  "KARAMEL PROD",
+  "KDFP",
+  "KELASSI ' SCHOOL",
+  "KEY FORM & SOLUTIONS",
+  "KEYCE ACADEMY - COLLEGE DE PARIS",
+  "KIWI INSTITUTE",
+  "KMK CONSEIL ET FORMATION",
+  "KONEXIO",
+  "KOUT QUE KOUT",
+  "KYAD CENTER",
+  "L ATELIER DES CHEFS",
+  "L INSTITUT DES AMERIQUES",
+  "L P PRIVE PIERRE DE COUBERTIN",
+  "L'ACADEMIE BY MG",
+  "L'ACADEMIE DES SERVICES AUX FAMILLES",
+  "L'ACADEMIE OUI CARE",
+  "L'ADAPT",
+  "L'ECOLE DES NOUVEAUX MONDES",
+  "L'ECOLE FRANCAISE",
+  "L'ECOLE NUMERIQUE",
+  "L'ECOLE ORCOM",
+  "L.A.C",
+  "L2B CONSEIL & FORMATION",
+  "LA BATISSE",
+  "LA BOX ACADEMIE",
+  "LA CITE DES FORMATIONS",
+  "LA COMPAGNIE DE FORMATION",
+  "LA FACTORY",
+  "LA FORBINE LYCEE TECHNIQUE PRIVE DE CHIMIE ET BIOLOGIE",
+  "LA GENERALE DES VEGETAUX",
+  "LA GRANDE CLASSE",
+  "LA PISCINE",
+  "LA PLATEFORME FORMATION",
+  "LA PRADETTE",
+  "LA SOURCE",
+  "LAUREAT FORMA",
+  "LB DEVELOPPEMENT",
+  "LB DEVELOPPEMENT DOM",
+  "LCMG CONSULTING",
+  "LCS INTERNATIONAL SAS",
+  "LE 101",
+  "LE BIHAN ACADEMIE (LBA)",
+  "LE CENTRE DES APPRENTIS D'ALSACE",
+  "LE CHIFFRE",
+  "LE PORTRAIT",
+  "LE STUDIO",
+  "LEADER ACADEMY",
+  "LEARN&FIT",
+  "LECOLE",
+  "LEGTA IND.LAITIERE AGRO-ALIMENTAIRE",
+  "LEPAFP",
+  "LES CLES DE L'ATELIER",
+  "LES FORMATIONS DU MARAIS",
+  "LES GARDERIEFORM",
+  "LES SOMMELIERS CAVISTES",
+  "LEVEL UP GROUP",
+  "LH COMMUNICATION",
+  "LIBERTE FORMATION",
+  "LILI ET COMPAGNIE",
+  "LIMOGES BUSINESS SCHOOL",
+  "LINEANE",
+  "LINK RH FORMATIONS & DEVELOPPEMENT",
+  "LN CONSEIL & CO",
+  "LO'R",
+  "LODIMA OUEST",
+  "LOGIC",
+  "LOHANDA",
+  "LOHATODE*/ISAAC  /",
+  "LOISIRS EDUCATION & CITOYENNETE FORMATION",
+  "LOUKILI",
+  "LRDTP",
+  "LS FORMATION",
+  "LSG ACADEMIE",
+  "LUBRANO",
+  "LUZCARE",
+  "LV CONSULTANTS",
+  "LY FORMATION",
+  "LYCEE DES METIERS PIERRE MENDES FRANCE",
+  "LYCEE ELLA FITZGERALD",
+  "LYCEE ENSEIGNEMENT AGRC PRIVE FONTLONGUE",
+  "LYCEE GEN & TECHNO L. ARMAND",
+  "LYCEE GENERAL ET TECHNOLOGIQUE CAMILLE JULLIAN",
+  "LYCEE GENERAL ET TECHNOLOGIQUE DOMINIQUE VILLARS",
+  "LYCEE GENERAL ET TECHNOLOGIQUE LIVET",
+  "LYCEE GENERAL ET TECHNOLOGIQUE PABLO PICASSO",
+  "LYCEE LIBERGIER",
+  "LYCEE POLYVALENT ANDRE MALRAUX",
+  "LYCEE POLYVALENT BOURDELLE",
+  "LYCEE POLYVALENT CHARLES ET ADRIEN DUPUY",
+  "LYCEE POLYVALENT CHEVROLLIER",
+  "LYCEE POLYVALENT ETAT LAETITIA BONAPARTE",
+  "LYCEE POLYVALENT JEAN BAPTISTE COLBERT",
+  "LYCEE POLYVALENT JEAN MERMOZ",
+  "LYCEE POLYVALENT JEAN ZAY",
+  "LYCEE POLYVALENT PAUL MATHOU",
+  "LYCEE POLYVALENT PHILIPPE DE GIRARD",
+  "LYCEE POLYVALENT TECHNOLOGIQUE E BELIN",
+  "LYCEE POLYVALENT VAUVENARGUES",
+  "LYCEE POLYVALENT XAVIER MARMIER",
+  "LYCEE PRIVE DE RIGNAC AMENAGEMENT PAYSAGER ELEVAGE BOVIN EQUIN SERVICES AUX PERS",
+  "LYCEE PROFESSIONNEL AGRICOLE HORTICOLE",
+  "LYCEE PROFESSIONNEL PRIVE BEL-AIR",
+  "LYCEE RAYMOND POINCARE",
+  "LYCEE TECHNOLOGIQUE-ECOLE NAT COMMERCE",
+  "LYCEE VITICOLE DE LA CHAMPAGNE",
+  "LYNXTITUT",
+  "LYON OLYMPIQUE UNIVERSITAIRE LOU RUGBY",
+  "M.PARTN'HAIR",
+  "M2A CONSULTING",
+  "M2I SCRIBTEL",
+  "M2S FORMATION",
+  "M2S FORMATION AUBAGNE SAS",
+  "M2S33 FORMATIONS",
+  "MAIS FAMIL URBAINE DE RENNES ST GREGOIRE",
+  "MAIS FAMILIALE RUR EDUCATION ORIENTATION",
+  "MAISON DE LA PROMOTION SOCIALE FORMATION",
+  "MAISON FAM RURALE EDUCATION ORIENTATION",
+  "MAISON FAM.RURALE LE VILLAGE",
+  "MAISON FAMI RURALE EDUCATION ORIENTATION",
+  "MAISON FAMIL EDUCAT ORIENTAT",
+  "MAISON FAMIL RURAL EDUCATION ORIENTATION",
+  "MAISON FAMIL RURALE EDUCAT ORIENT",
+  "MAISON FAMIL RURALE EDUCATION ET ORIENTA",
+  "MAISON FAMILIALE",
+  "MAISON FAMILIALE DE LA ROQUE D ANTHERON",
+  "MAISON FAMILIALE DE VOUGY",
+  "MAISON FAMILIALE DU LIBOURNAIS",
+  "MAISON FAMILIALE HORTICOLE",
+  "MAISON FAMILIALE LES EBEAUX",
+  "MAISON FAMILIALE RURALE",
+  "MAISON FAMILIALE RURALE D EDUCATION",
+  "MAISON FAMILIALE RURALE D EDUCATION ET D ORIENTATION DE ANSE",
+  "MAISON FAMILIALE RURALE D EDUCATION ET D ORIENTATION DE SALLANCHES LE BELVEDERE",
+  "MAISON FAMILIALE RURALE D'ANNECY LE VIEUX",
+  "MAISON FAMILIALE RURALE D'EDUCATION ET D'ORIENTATION DE COUBLEVIE",
+  "MAISON FAMILIALE RURALE D'EDUCATION ET D'ORIENTATION DU BLAYAIS",
+  "MAISON FAMILIALE RURALE DE LA PLAINE DES PALMISTES",
+  "MAISON FAMILIALE RURALE DE SAINT ETIENNE",
+  "MAISON FAMILIALE RURALE DES BARONNIES",
+  "MAISON FAMILIALE RURALE DU RIBERACOIS",
+  "MAISON FAMILIALE RURALE DU VAL DE MANSE",
+  "MAISON FAMILIALE RURALE EDUC ET ORIENTAT",
+  "MAISON FAMILIALE RURALE EDUCA ORIEN",
+  "MAISON FAMILIALE RURALE EDUCA ORIENTA",
+  "MAISON FAMILIALE RURALE EDUCAT ET ORIENT",
+  "MAISON FAMILIALE RURALE EDUCAT ORIENTAT",
+  "MAISON FAMILIALE RURALE EDUCATION  CAMPUS DU MARTELET",
+  "MAISON FAMILIALE RURALE EDUCATION ORIENT",
+  "MAISON FAMILIALE RURALE ET D'EDUCATION",
+  "MAISON FAMILIALE RURALE LA PALMA",
+  "MAISON FAMILIALE RURALE MIXTE",
+  "MAISON FAMILIALE RURALE ORGANISME DE FORMATION PAR APPRENTISSAGE DE TALMONT",
+  "MAISON FAMILIALE RURALE PETITE CAMARGUE",
+  "MAJY INSERTION",
+  "MASSENO",
+  "MAUFFREY ACADEMY",
+  "MAYOTTE CONSULTING ET FORMATION",
+  "MBA INSTITUTE",
+  "MCM FORMATION",
+  "MDA CONSULTING",
+  "MEDEO FORMATION",
+  "MEDIA FORMATION",
+  "MENDIBOURE FORMATION",
+  "MENSA FORMATION",
+  "MENTOR INSTITUT",
+  "METADXS",
+  "METROPOLE D'AIX-MARSEILLE-PROVENCE",
+  "METROPOLE NICE COTE D AZUR",
+  "MEWO",
+  "MFR DE L ENTRE 2 MERS",
+  "MFR DE RUMENGOL",
+  "MFR LA MONTAGNE",
+  "MFR LA PINEDE",
+  "MFR SUD AGROMAT-CAMPUS ALTERNANCE",
+  "MICHEL CONDOMITTI",
+  "MINDSHIFT GROUPE",
+  "MIRABILIS",
+  "MK FORMATION",
+  "MODULA FORMATION",
+  "MRH ANTILLES",
+  "MTV",
+  "MVP",
+  "NACARAT",
+  "NANTES UNIVERSITE",
+  "NAS",
+  "NEO",
+  "NEO SPHERE",
+  "NEOGEST EDUCATION GROUP",
+  "NEPSOD EVOLUTION",
+  "NESS EDUCATION - FMS -SMBS-FBS",
+  "NEWAY",
+  "NEXT-U FRANCE",
+  "NEXA",
+  "NEXA DIGITAL SCHOOL",
+  "NOMAD FORMATION",
+  "NORMANDIE FORMATION PROPRETE",
+  "NOSCHOOL",
+  "NOUAS FORMATION",
+  "NOUVELLE ASSOC EDUCAT POPULAI ST-BENIGNE",
+  "NOVEI",
+  "NOVETUDE SANTE PRO",
+  "NUEVO ENTREPRISE",
+  "O.G. CENTRE GUSTAVE EIFFEL",
+  "OBJECTIF COMPETENCES",
+  "OCF-EST",
+  "OCTO",
+  "OFFICE E-FORMA",
+  "OFPRO IMMO",
+  "OGEC CLAUDE DAUNOT - SAINT LEON IX",
+  "OGEC DE CHARLEVILLE MEZIERES",
+  "OGEC GROUPE SAINT JOSEPH LA SALLE - DIJON",
+  "OGEC LA JOLIVERIE",
+  "OGEC LES CHARMILLES",
+  "OGEC LTP ST JEAN BAPTISTE DE LA SALLE",
+  "OGEC LYCEES ET INSTITUT MONTPLAISIR",
+  "OGEC NANTES ERDRE",
+  "OGEC SAINT MICHEL",
+  "OGEC STE CATHERINE",
+  "OGEC VINCENT DE PAUL",
+  "OGEC-FONDATION DON BOSCO",
+  "OIF REUNION",
+  "OKAN FORMATION",
+  "OKTOGONE",
+  "OMNIS",
+  "ONE 2 ONE",
+  "ONE BUSINESS SCHOOL",
+  "ONLINEFORMAPRO",
+  "OPEN BAR",
+  "OPENCLASSROOMS",
+  "OPTIMA",
+  "OPUS FORMATION",
+  "ORDIA",
+  "ORFEE",
+  "ORG DE GEST DE L'ENSEIGNEMT CATHO DU LYCEE LEVAVASSEUR ET DE SON ETABLISSEMENT SECONDAIRE CFA-ECR",
+  "ORGALY",
+  "ORGALY FRANCE",
+  "ORGALY MARSEILLE",
+  "ORGANI GEST INSTIT CATHO ETUDE SUPER",
+  "ORGANISM GESTION ENSEIGNEMENT CATHOLIQUE",
+  "ORGANISME DE FORMATION POUR L'INSERTION L'ACCOMPAGNEMENT ET LA QUALIFICATION",
+  "ORGANISME DE GESTION DE L'ECOLE SCHNEIDER ELECTRIC",
+  "ORGANISME DE GESTION DE L'ENSEIGNMENT CATHOLIQUE EMMANUEL D'ALZON",
+  "ORLEANS METROPOLE",
+  "OSENGO",
+  "OSEZ ENTREPRENDRE",
+  "OXIDEVE",
+  "OXYTALIS",
+  "OZAR FORMATION",
+  "P.A.C.T. AMAZONE",
+  "PA-COURS VERSAILLAIS-IGECOM SUP'",
+  "PAIE & RH ACADEMIE",
+  "PALAFIS",
+  "PAM OI",
+  "PARIS ECOLE DE MANAGEMENT",
+  "PARIS NORD BUSINESS SCHOOL",
+  "PARIS YNOV CAMPUS",
+  "PASSEPORT PREVENTION",
+  "PCWORKS31",
+  "PERFORM COIFFURES",
+  "PERFORMA",
+  "PHILIANCE",
+  "PIGIER",
+  "PIKANGO",
+  "PLATE FORME DE RENCONTRES, D'ORIENTATION ET DE FORMATIONS INTER- REGIONALE DE L'IMPRESSION DU LIVRE ET DE L'EDITION",
+  "PLATINIUM ACADEMY",
+  "POLE DE FORMATION PASTEUR",
+  "POLE ESSONNE RESSOURCES FORMATIONS",
+  "POLE ETUDE RECHERCHE FORMATION",
+  "POLE FORMATION DES INDUSTRIES TECHNOLOGIQUES",
+  "POLYVIA FORMATION",
+  "PRASMES",
+  "PRECELLENCE FORMATIONS ET CONSEILS",
+  "PREP'AVENIR FORMATION",
+  "PREPA FORMATION",
+  "PREVENTIS ANTILLES GUYANE",
+  "PRISMA FORMATIONS",
+  "PRIVATE FRENCH UNIVERSITY",
+  "PRO FOR TECH",
+  "PRO FORMAT",
+  "PRO S.A.P. FORMATIONS",
+  "PRO SYSTEMES",
+  "PRO-ALTERNA",
+  "PRO-FYL",
+  "PROACTIVE ACADEMY",
+  "PROFESSION SPORTS DANS LE PAS DE CALAIS",
+  "PROFIL MANAGER",
+  "PROGRESS COM",
+  "PROGRESS SUP",
+  "PROJET PROFESSIONNEL PLUS",
+  "PROMEO CENTRE DE FORMATION D APPRENTIS DE L INDUSTRIE DE PICARDIE PROMEO CFAI PICARDIE",
+  "PROMOTRANS FORMATION PROFESSIONNELLE CONTINUE",
+  "PROPULS'UP",
+  "PROVENCE FORMATION",
+  "PSAC",
+  "PSBS",
+  "PSTB-PARIS SCHOOL OF TECHNOLOGY & BUSINESS EN ABREGE P.S.T.B",
+  "PSTM",
+  "PSXR",
+  "PURPLE CAMPUS",
+  "QUALIFORMATION",
+  "QUALITY FORMATION",
+  "QWANTIC",
+  "RAISE ON",
+  "RE FORMATION",
+  "REBONDIR",
+  "RECRUTEMENT,FORMATION,CONSULTING ET INGENIERIE",
+  "RETRAVAILLER DANS L OUEST",
+  "REUNI'FORM",
+  "REUNION APPRENTISSAGE FORMATION",
+  "REUSSIR",
+  "RHREFLEX",
+  "RI7",
+  "RIG UP FORMATION",
+  "ROCKET SCHOOL",
+  "ROCKET SCHOOL LYON",
+  "ROYER ROBIN ASSOCIES",
+  "S-TEAM FORMATIONS",
+  "S2M FORMATION",
+  "SA SOFTEC AVENIR FORMATION",
+  "SABEXTRA",
+  "SAG&FORM",
+  "SAINT DOMINIQUE",
+  "SALLTO - SIPCA",
+  "SAM & FT ACCOMPAGNEMENT",
+  "SANTEST",
+  "SARL AVENIR FORMATION CONSEIL",
+  "SARL COACHING PACA",
+  "SARL FR'HAIR",
+  "SARL LES SCRIBES",
+  "SARL RESTHO FORMATION",
+  "SAS 3 B CONSEIL",
+  "SAS E.C.D.E.",
+  "SAS L'ACADEMIE DE MANAGEMENT",
+  "SAS LE ROSEAU CONSEIL",
+  "SAS MOLERA FORMATION",
+  "SASU CWT, DEVELOP YOUR TALENT",
+  "SASU DEFI",
+  "SASU ELFE COACHING CONSEIL ET ASSISTANCE",
+  "SASU IROF (INSTITUT REUNIONNAIS D'ORIENTATION ET DE FORMATION)",
+  "SAUVEGARDE 13",
+  "SAVOIRS ET FORMATION",
+  "SBS GROUPE",
+  "SCHOLA.2",
+  "SCHOLAR FAB ORGANISATION",
+  "SCHOLIA",
+  "SCHOOL EFMDN BEAUTY",
+  "SCHOOL OF IMPACT",
+  "SCIENCES-U LILLE",
+  "SCOP INSTEP",
+  "SCOTTO DI CESARE L AUTRE ECOLE COIFFURE",
+  "SE FORMER EN PAIE",
+  "SERPE FORMATION",
+  "SESAM FORMATIONS",
+  "SH FORMATION",
+  "SIAF FORMATION",
+  "SIGMA FORMATION",
+  "SILVYA TERRADE GRAND-EST",
+  "SILVYA TERRADE GRAND-OUEST",
+  "SILVYA TERRADE PARIS CHAMPS ELYSEES",
+  "SILVYA TERRADE SUD-EST",
+  "SILVYA TERRADE SUD-OUEST",
+  "SIMPLON.CO",
+  "SINEQUANONE CONSEIL",
+  "SKALA BUSINESS SCHOOL",
+  "SKILL AND YOU",
+  "SKILLS DEVELOPMENT",
+  "SKILLS MAJOR",
+  "SKILLZFORMATION",
+  "SKOOL N'JOB",
+  "SOC EUROPEENNE DE FORMATION",
+  "SOC FORMATION EMPLOI INSERTION",
+  "SOCIETE CIFEP SASU",
+  "SOCIETE DEBIEU",
+  "SOCIETE INSTITUT TECHNIQUE ET PROFESSIONNEL AIXOIS",
+  "SOCIETE MAHORAISE DE FORMATION PROFESSIONNELLE",
+  "SOCOTEC CAMPUS",
+  "SOGESTE",
+  "SOLARIS DEVELOPPEMENT",
+  "SOLUTYS",
+  "SONELO",
+  "SOURCE DE BEAUTE FORMATION",
+  "SPE FORMATION",
+  "SPECTACLES ET TECHNI ASS FRANC FORMATION",
+  "SPORT 4.0",
+  "SPORTS & ETUDES CONCEPT",
+  "STADE FORMATION",
+  "STAJ AURA / SERVICE TECHNIQUE POUR LES ACTIVITES DE JEUNESSE AUVERGNE RHONE ALPES",
+  "STAND UP FORMATION",
+  "STE D'EXPLOITATION RUFFEL BEZIERS",
+  "STELLA DI L ALBA",
+  "STELO FORMATION",
+  "STEPHENSON FORMATION",
+  "STK",
+  "STOP AU CHOMAGE",
+  "STUDI",
+  "SUD CAMPUS",
+  "SUD ERKA",
+  "SUD FORMATION CONSEIL",
+  "SUD MANAGEMENT",
+  "SUD PREVENTION SECURITE",
+  "SUP DES RH",
+  "SUP' COURS PERET",
+  "SUP'BUSINESS SCHOOL OF MANAGEMENT",
+  "SUP'EXPERTISE",
+  "SUPEXUP CONSULTANT",
+  "SUSTAINABLE DEVELOPMENT MANAGEMENT INSTITUTE ET INSTITUT DE MANAGEMENT DU DEVELOPPEMENT DURABLE",
+  "SYMPHONIE ACCOMPAGNEMENTS",
+  "SYNDICAT MIXTE POUR LA GESTION DU CENTRE DE FORMATION D'APPRENTIS DE L'EST DU LOIRET",
+  "SYNERGIE OCEAN INDIEN",
+  "SYNETHIS",
+  "T.M.K. PERFORMANCES",
+  "TALIS B.S. PARIS",
+  "TALIS BAYONNE",
+  "TDMR",
+  "TEDDY PARRA SAS",
+  "TEMANA",
+  "TEMPUS FORMATIONS",
+  "TERA CAMPUS",
+  "TETRANERGY",
+  "THALASS ' SUN FORMATION",
+  "THE BEAUTY TRAIN",
+  "THE BRIDGE ECOLE ENTREPRISES",
+  "THE CODE",
+  "THE INFLUENCER HIGH SCHOOL",
+  "THE NUUM FACTORY",
+  "THE WELCOME FACTORY",
+  "THOT LEARNING",
+  "TI LAMPE",
+  "TKL FORMA",
+  "TOCCATA",
+  "TONYE'S FASHION ACADEMY",
+  "TOTEM INITIAL",
+  "TRANS-FAIRE",
+  "TREMPLIN 84",
+  "TROIS SOLUTIONS POUR L'ALTERNANCE",
+  "TROYES AUBE FORMATION",
+  "UCPA FORMATION",
+  "UDD",
+  "UNIFADOM",
+  "UNION NATION RURALE EDUCATION PROMOT EST",
+  "UNION NATIONAL INDUS CARRIERE MATERIEL",
+  "UNION NATIONALE DES FLEURISTES",
+  "UNIV PARIS XIII PARIS-NORD VILLETANEUSE",
+  "UNIVERSITE D ARTOIS",
+  "UNIVERSITE D'ANGERS",
+  "UNIVERSITE D'EVRY VAL D'ESSONNE",
+  "UNIVERSITE DE BRETAGNE SUD",
+  "UNIVERSITE DE CAEN NORMANDIE",
+  "UNIVERSITE DE GUYANE",
+  "UNIVERSITE DE LA REUNION",
+  "UNIVERSITE DE LIMOGES",
+  "UNIVERSITE DE MONTPELLIER",
+  "UNIVERSITE DE POITIERS",
+  "UNIVERSITE DE REIMS CHAMPAGNE-ARDENNE",
+  "UNIVERSITE DE RENNES",
+  "UNIVERSITE DE ROUEN-NORMANDIE",
+  "UNIVERSITE DE TECHNOLOGIE DE COMPIEGNE",
+  "UNIVERSITE DE TECHNOLOGIE DE TARBES",
+  "UNIVERSITE DES COMPETENCES HABITAT",
+  "UNIVERSITE DU LITTORAL COTE D'OPALE",
+  "UNIVERSITE DU MANS",
+  "UNIVERSITE GRENOBLE ALPES",
+  "UNIVERSITE GUSTAVE EIFFEL",
+  "UNIVERSITE JEAN MONNET SAINT ETIENNE",
+  "UNIVERSITE LYON 2 A ET L LUMIERE",
+  "UNIVERSITE PARIS EST CRETEIL VAL DE MARNE",
+  "UNIVERSITE PARIS NANTERRE",
+  "UNIVERSITE RENNES  II  HAUTE BRETAGNE",
+  "UNIVERSITE TOULOUSE III - PAUL SABATIER",
+  "UPYZI",
+  "VAELIA",
+  "VAL BUSINESS SCHOOL",
+  "VALEUR AJOUTEE",
+  "VIA FORMATION",
+  "VIATICUS PARDASA",
+  "VICTORIA",
+  "VOL GROUPE CONSEILS, FORMATIONS",
+  "VOLTAIRE BUSINESS SCHOOL",
+  "VOLTEE",
+  "VSM CFA",
+  "WATERFORM",
+  "WEB ASSOCIATION BERGERAC",
+  "WENO FORMATION",
+  "WIKI LEARN FORMATION",
+  "WIN COMPETENCE FORMATION",
+  "WKG CORPORATE",
+  "XP SCHOOL",
+  "YAFFA",
+  "YOUSCHOOL",
+  "YOU WEB",
+  "ZILLENIUM",
+  "ZLS",
+  "ZONE 01 ROUEN",
+]
+
+const stringNormaliser = (str: string): string => {
+  return removeAccents(str.toLowerCase())
+}
+
+const normalizedCfaList = cfaCompanyList.map(stringNormaliser)
+
+export const isCompanyInBlockedCfaList = (nom: string | null | undefined): boolean => {
+  if (!nom) return false
+  const nomNormalise = stringNormaliser(nom)
+  return normalizedCfaList.includes(nomNormalise)
+}
